requirements.txt
pillow
streamlit
ultralytics